+++ core/src/com/shapeshifter/Actor/MovementStrategy/UserMovementStrategy.java
@@ -20,7 +20,7 @@ public class UserMovementStrategy implements MovementStrategy{
     }
 
     @Override
-    public float getNewAngularSpeed() {
+    public double getNewAngularSpeed() {
         if (Gdx.input.isKeyPressed(Input.Keys.A)) return source.getMaxAngularSpeed();
         if (Gdx.input.isKeyPressed(Input.Keys.D)) return -source.getMaxAngularSpeed();
         return 0;
